Removed dependency for our Skybrud.LinkPicker package
Since we only used a few components from the link picker package, those components have now been copied so they're avaiable directly in the redirects package as well. This means that we can drop Skybrud.LinkPicker as a dependency, which also means that Skybrud.Umbraco.GridData (a dependency of Skybrud.LinkPicker) won't be installed with the redirects package.