Fixed an issue with redirect type not being properly set
If a redirect was created from the global dashboard, the type/mode would always be set to "content", which obviously is wrong. The redirects would still work fine, but eg. a redirect for a media (created via the dashboard), wouldn't show up in the redirects proeprty editor of said media due to the wrong type being stored in the database.