Added new Editor View and Controller for picking the root node for redirects.
This allows users to create site specific redirects in Umbraco 8 multi site instances. Most of the work was done in the front end as the backend code was already set up to work with a root node Id. All I had to do in the backend was to update the AddRedirectOptions and EditRedirectOptions classes and remove the JsonIgnore attribute off RootNodeId and replace it with a JsonProperty attribute. I also had to wire up RootNodeId on EditRedirectOptions to be used on save when editing a redirect.