Rewrote migration to work with SQL CE
The migration was based on a SQL statement using a nested query, which worked fine on a regular SQL Server, but not so much on SQL CE as nested queries apparently isn't supported. With this commit, the migration will instead fetched all affected rows, and then update each row individually in the database.