Event should allow writing opposed to just reading
The event logic has now been updated so setting a redirect from one of the two events is now supported. This means that if a redirect is set from the pre lookup event, the redirect will not attempt to do a lookup in the database. The post lookup can now also be used to either set or overwrite the redirect found by this package.

The HTTP module has also been updated so that the post lookup event is invoked before the local "redirectUrl" variable is initialized. This means that if the post lookup event sets a different redirect, the destination of that redirect will be used instead.