Introduced new "RedirectsSearchOptions" class
The "GetRedirects" method in the redirects service now takes an instance of "RedirectsSearchOptions" instead of the individual parameters. This for one cleans up the code a bit, but also allows us to add more parameters/options to the "RedirectsSearchOptions" class at a later point, without changing the signature of the "GetRedirects" method, thus avoinding breaking changes.