Updated migration with better approach for dropping column
Fixes #106

Older versions of SQL Server appear to not support the "IF EXISTS" syntax used when attempting to drop the IsRegex column from the redirects table, resulting in issues when installing this package. With this commit, the migration uses Umbraco's logic to first check whether the column does in fact exist, and the migration will drop the column if it does.