Fixed incorrect destination warning
The redirects dashboard shows a warning if the destination is content or media, and the destination is either unpublished, trashed or deleted.

The dashboard would incorrectly show an unpublished warning for redirects have been deleted. With this commit, the correct warning is now being returned by the underlying API, so the dashboard doesn't have to check other properties to determine whether a warning should be shown, and if so, what it should show.