Updated how the package is configured during installation
If the redirects package was installed before the upgrade installer was finished, two exceptions would occur:

- Migration to install the database would fail as the database isn't during before the Umbraco installer has been completed.
- Middleware would also run when trying to access the Umbraco installer, thus failing as the middleware would try to access the database through the redirects service

With this commit, the migration plan for the redirects package will only be set up when the runtime level is "Run".

The middleware will now not run when the runtime level is "Install", but will still run when the runtime level is "Upgrade". This might need a bit more testing, but seems to be fine as there *is* a database at this level.