Added missing ForwardQueryString check to GetRedirectByPathAndQuery
GetRedirectByPathAndQuery(guid, string, string) appears to do the same thing as GetRedirectByUrl(guid, string, string) assuming some logic is unintentially missing, so added the fallback ForwardQueryString check. Consolidated the methods and marked GetRedirectByUrl(guid, string, string)  as obsolete.

That exposed no code calling the GetDestinationUrl which critically merges the query string parameters so added that to the middleware.

Appears to fix #43 and possibly #104