Introduced the "RedirectsServiceDependencies" and "RedirectsBackOfficeHelperDependencies" classes
As developers using this package might extend either the "RedirectsService" or the "RedirectsBackOfficeHelper" classes, it would be a breaking change if we later have to introduce new dependencies via the classes constructor. So to hopefully avoid this, this commit introduces the two new classes which will be used as "dependency intermediaries". So for instance if we later need an addtional dependency for the "RedirectsService", we can add the dependency in the "RedirectsServiceDependencies" class instead without introducing a breaking change.