The various "GetDestinationUrl" methods should not be nullable
Ideally the saved destination URL of a redirect should never be null or empty, so if this is the case, it indicates a malformed redirect. If all redirects are created through the UI or the package's default implementation for "IRedirectService", this shouldn't happen, so it should be okay to throw an exception instead of returning "null".