Stripped comment from "RedirectsPackage.InformationalVersion" property value
When using the compiler that comes with .NET 8, the hash of the current commit is automatically appended to the informational version. While technically a part of the information version, we don't really need/want it for the purposes of the "RedirectsPackage.InformationalVersion" property.

(cherry picked from commit 4e810e6a84ead2e790df62a4dd8222486f4e6948)