Additional adjustments to the query string forwarding logic
In earlier versions of this package the query string forwarding logic were primarily handled by querying the database, which was case insensitive. After moving some of the logic to C#, this part was somewhat unintentionally case sensitive, while the part still handled in the database was still case insensitive. With this commit both the database query and the C# logic are now case sensitive to better match the previous behavior.

Whether it should in fact be case insensitive is however a whole other question. It probably shouldn't be, but it has been like this ever since the first release as SQL WHERE clauses are case insensitive. So we shouldn't change this behavior now.

Related to #193

(cherry picked from commit 542a4a81447a6020e89e6d5f6b9e302766f98d4c)